# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
...
        "/comments": {
            "get": {
                "description": "Get a list of Comments.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comments"
                ],
                "summary": "Get all Comments.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Comments"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Creating a new Comments.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comments"
                ],
                "summary": "Create New Comments.",
                "parameters": [
                    {
                        "description": "the body to create a new Comments",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.commentsInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Comments"
                        }
                    }
                }
            }
        },
        "/comments/{id}": {
            "get": {
                "description": "Get an Comments by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comments"
                ],
                "summary": "Get Comments.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comments id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Comments"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Delete a Comments by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comments"
                ],
                "summary": "Delete one Comments.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comments id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Update Comments by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comments"
                ],
                "summary": "Update Comments.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comments id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "the body to update comments",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.commentsInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Comments"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.commentsInput": {
            "type": "object",
            "properties": {
                "comments_section": {
                    "type": "string"
                },
                "devices_id": {
                    "type": "integer"
                },
                "reviews_id": {
                    "type": "integer"
                },
                "your_name": {
                    "type": "string"
                }
            }
        },
...
        "models.Comments": {
            "type": "object",
            "properties": {
                "comments_section": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "devicesID": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "reviewsID": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "your_name": {
                    "type": "string"
                }
            }
        },
...
    }
}